Create app state to hold settings panel info

+++ src/app/middleware/persistMiddleware.js
@@ -1,4 +1,4 @@
-import { denormalize, schema } from "normalizr";
+import { denormalize, schema } from 'normalizr';
 
 // Persist the board to the database after almost every action.
 const persistMiddleware = store => next => action => {
@@ -13,49 +13,54 @@ const persistMiddleware = store => next => action => {
 
   // Nothing is persisted for guest users
   if (user) {
-    if (action.type === "DELETE_BOARD") {
-      fetch("/api/board", {
-        method: "DELETE",
+    if (
+      action.type === 'TOGGLE_SETTINGS_MENU' ||
+      action.type === 'SETTINGS_PENDING'
+    )
+      return;
+    if (action.type === 'DELETE_BOARD') {
+      fetch('/api/board', {
+        method: 'DELETE',
         body: JSON.stringify({ boardId }),
-        headers: { "Content-Type": "application/json" },
-        credentials: "include"
+        headers: { 'Content-Type': 'application/json' },
+        credentials: 'include'
       });
-    } else if (action.type === "UPDATE_USER") {
-      const user2 = new schema.Entity("user", {}, { idAttribute: "_id" });
+    } else if (action.type === 'UPDATE_USER') {
+      const user2 = new schema.Entity('user', {}, { idAttribute: '_id' });
       const entities = { user };
       const userData = denormalize(user, user2, entities);
 
-      fetch("/api/user", {
-        method: "PUT",
+      fetch('/api/user', {
+        method: 'PUT',
         body: JSON.stringify(userData),
-        headers: { "Content-Type": "application/json" },
-        credentials: "include"
+        headers: { 'Content-Type': 'application/json' },
+        credentials: 'include'
       });
       // All action-types that are not DELETE_BOARD or PUT_BOARD_ID_IN_REDUX are currently modifying a board in a way that should
       // be persisted to db. If other types of actions are added, this logic will get unwieldy.
-    } else if (action.type !== "PUT_BOARD_ID_IN_REDUX") {
+    } else if (action.type !== 'PUT_BOARD_ID_IN_REDUX') {
       // Transform the flattened board state structure into the tree-shaped structure that the db uses.
-      const card = new schema.Entity("cardsById", {}, { idAttribute: "_id" });
+      const card = new schema.Entity('cardsById', {}, { idAttribute: '_id' });
       const list = new schema.Entity(
-        "listsById",
+        'listsById',
         { cards: [card] },
-        { idAttribute: "_id" }
+        { idAttribute: '_id' }
       );
       const board = new schema.Entity(
-        "boardsById",
+        'boardsById',
         { lists: [list] },
-        { idAttribute: "_id" }
+        { idAttribute: '_id' }
       );
       const entities = { cardsById, listsById, boardsById };
 
       const boardData = denormalize(boardId, board, entities);
 
       // TODO: Provide warning message to user when put request doesn't work for whatever reason
-      fetch("/api/board", {
-        method: "PUT",
+      fetch('/api/board', {
+        method: 'PUT',
         body: JSON.stringify(boardData),
-        headers: { "Content-Type": "application/json" },
-        credentials: "include"
+        headers: { 'Content-Type': 'application/json' },
+        credentials: 'include'
       });
     }
   }
